+++ experiments/experiments.go
@@ -18,6 +18,7 @@ var (
 	GentleForce     Experiment
 	RemoteTaskfiles Experiment
 	EnvPrecedence   Experiment
+	Plugins         Experiment
 )
 
 // Inactive experiments. These are experiments that cannot be enabled, but are
@@ -45,6 +46,7 @@ func Parse(dir string) {
 	GentleForce = New("GENTLE_FORCE", config, 1)
 	RemoteTaskfiles = New("REMOTE_TASKFILES", config, 1)
 	EnvPrecedence = New("ENV_PRECEDENCE", config, 1)
+	Plugins = New("PLUGINS", config, 1)
 	AnyVariables = New("ANY_VARIABLES", config)
 	MapVariables = New("MAP_VARIABLES", config)
 }