+++ setup.go
@@ -13,6 +13,7 @@ import (
 	"github.com/sajari/fuzzy"
 
 	"github.com/go-task/task/v3/errors"
+	"github.com/go-task/task/v3/experiments"
 	"github.com/go-task/task/v3/internal/env"
 	"github.com/go-task/task/v3/internal/execext"
 	"github.com/go-task/task/v3/internal/filepathext"
@@ -82,6 +83,11 @@ func (e *Executor) readTaskfile(node taskfile.Node) error {
 		taskfile.WithPromptFunc(promptFunc),
 	)
 	graph, err := reader.Read(ctx, node)
+	if experiments.Plugins.Enabled() {
+		if err := reader.LoadPlugin(ctx, node); err != nil {
+			return err
+		}
+	}
 	if err != nil {
 		if errors.Is(err, context.DeadlineExceeded) {
 			return &errors.TaskfileNetworkTimeoutError{URI: node.Location(), Timeout: e.Timeout}